Cleanup driverstation code
This commit reorganizes the driverstation code. The main thing it does
is organize the field socket handling into a class called
FieldConnnection. Eventually, this will be on its own thread with the
GUI on a separate thread.

In addition, this commit adds a driverstation
class to manage the driverstation application
(basically removing the code from the main method).